requirements.txt
asgiref==3.7.2
attrs==23.1.0
coverage==7.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.11.1
django-auditlog==2.0.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-import-export==3.2.0
django-jsonfield==1.4.1
django-reorder-admin==0.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
et-xmlfile==1.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupPy==1.14
model-bakery==1.15.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pluggy==1.3.0
psycopg==3.1.10
pycodestyle==2.11.0
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.6.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
